+++ packages/flutter_tools/lib/src/extension/doctor.dart
