+++ packages/flutter_tools/pubspec.yaml
@@ -119,4 +119,4 @@ dartdoc:
   # Exclude this package from the hosted API docs.
   nodoc: true
 
-# PUBSPEC CHECKSUM: a631
+# PUBSPEC CHECKSUM: 5030
